% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{actinn}
\alias{actinn}
\title{------------------------
---Supervised methods---
------------------------}
\usage{
actinn(train, test, label_train, python_link = NULL)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{python_link}{The absolute path of python that tensorflow 1 is installed.}
}
\value{
The predict cell type.
}
\description{
ACTINN method
The tensorflow1 is used in ACTINN, please set the right python version in R through reticulate::use_python().
}
