% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{castle}
\alias{castle}
\title{CaSTLe method}
\usage{
castle(train, test, label_train, threshold = 0.7)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{threshold}{threshold for 'Unknown' cell type.}
}
\value{
The predict cell type.
}
\description{
CaSTLe method
}
