% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{cellblast}
\alias{cellblast}
\title{CELLBLAST method}
\usage{
cellblast(train, test, label_train, python_link = NULL)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}
}
\value{
The predict cell type.
}
\description{
CELLBLAST method
}
