% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{cellfishing.ji}
\alias{cellfishing.ji}
\title{Cellfishing.ji method}
\usage{
cellfishing.ji(train, test, label_train, k = 10)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{k}{a positive integer specifying the number of nearest neighbours to find.}
}
\value{
The predict cell type.
}
\description{
Cellfishing.ji method
}
