% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{clustifyr}
\alias{clustifyr}
\title{clustifyR method}
\usage{
clustifyr(
  train,
  test,
  label_train,
  time = F,
  min.cells = 0,
  min.features = 0,
  normalization.method = "LogNormalize",
  scale.factor = 10000,
  selection.method = "vst",
  nfeatures = 2000,
  mean.cutoff = c(0.1, 8),
  dispersion.cutoff = c(1, Inf),
  features = NULL,
  vars.to.regress = NULL,
  npcs = 50,
  k.param = 20,
  dims = 1:10,
  resolution = 0.8
)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{time}{return computation time, default is FALSE.}

\item{min.cells}{Include features detected in at least this many cells. Will subset the counts matrix as well. To reintroduce excluded features, create a new object with a lower cutoff. Default: 0.}

\item{min.features}{Include cells where at least this many features are detected. Default: 0.}

\item{normalization.method}{Method for normalization. Default: LogNormalize.}

\item{scale.factor}{Sets the scale factor for cell-level normalization. Default: 10000.}

\item{selection.method}{How to choose top variable features. Default: vst.}

\item{nfeatures}{Number of features to select as top variable features; only used when selection.method is set to 'dispersion' or 'vst'. Default: 2000.}

\item{mean.cutoff}{A two-length numeric vector with low- and high-cutoffs for feature means. Default: c(0.1, 8).}

\item{dispersion.cutoff}{A two-length numeric vector with low- and high-cutoffs for feature dispersions. Default: c(1, Inf).}

\item{features}{Vector of features names to scale/center. Default is variable features. Default: NULL.}

\item{vars.to.regress}{Variables to regress out (previously latent.vars in RegressOut). For example, nUMI, or percent.mito. Default: NULL.}

\item{npcs}{Total Number of PCs to compute and store (50 by default).}

\item{k.param}{Defines k for the k-nearest neighbor algorithm. Default: 20.}

\item{dims}{Dimensions of reduction to use as input. Default: 1:10.}

\item{resolution}{Value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities. Default: 0.8.}
}
\value{
The predict cell type.
}
\description{
clustifyR method
}
