% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{digitalcellsorter}
\alias{digitalcellsorter}
\title{-----------------------------
---Lazy-supervised methods---
-----------------------------
-------------------------------
---Marker-supervised methods---
-------------------------------}
\usage{
digitalcellsorter(test, marker_list = NULL, time = F)
}
\arguments{
\item{test}{Same format like train.}

\item{marker_list}{cell type marker list contains marker genes.}

\item{time}{return computation time, default is FALSE.}
}
\value{
The predict cell type.
}
\description{
DigitalCellSorter method
}
