% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{Evaluate}
\alias{Evaluate}
\title{Evaluate the prediction performance, source from PMID: 31500660.}
\usage{
Evaluate(ture_label, pred_label)
}
\arguments{
\item{ture_label}{A list contains the ture label of cell type.}

\item{pred_label}{A list contains the predict label of cell type.}
}
\value{
A list contaions the table info of ture/predict label, the F1 score of each cell type, the median F1 score predict label, the mean F1 score of
predict label, the accuracy rate of predict label,the percentage of 'Unknown' cell type, the table info of ture label.
}
\description{
Evaluate the prediction performance, source from PMID: 31500660.
}
