% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazySupervised.R
\name{lazysupervised}
\alias{lazysupervised}
\title{UnSupervised method to assign cell type}
\usage{
lazysupervised(train, test, label_train, method, ...)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{method}{Choosing the published method to assign cell type. Now support method contains: CELLBLAST, CellFishing.ji,scmapcell}

\item{...}{other paramter used in method/classifer, the detial could see the corresponding method.}
}
\value{
The predict cell type.
}
\description{
UnSupervised method to assign cell type
}
