% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{loadRData}
\alias{loadRData}
\title{load R object and assign new name.}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{File path of RData.}
}
\value{
R object.
}
\description{
load R object and assign new name.
}
