% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffMarker.R
\name{markerconvert}
\alias{markerconvert}
\title{marker list to matrix}
\usage{
markerconvert(markers)
}
\arguments{
\item{markers}{A list contains the marker gene of each cell type, the list name is cell type.}
}
\value{
A 0-1 matrix that row is marker gene, col is cell type.
}
\description{
marker list to matrix
}
