% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{markercount_semi}
\alias{markercount_semi}
\title{MarkerCount_semi method}
\usage{
markercount_semi(test, marker_list, time = F)
}
\arguments{
\item{test}{Same format like train.}

\item{marker_list}{A list contains marker gene of cell type.}

\item{time}{return computation time, default is FALSE.}
}
\value{
The predict cell type.
}
\description{
MarkerCount_semi method
}
