% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerSupervised.R
\name{markersupervised}
\alias{markersupervised}
\title{Semi-supervised method to assign cell type}
\usage{
markersupervised(test, marker_list, method, ...)
}
\arguments{
\item{test}{A gene-cell matrix of single cell expression data.}

\item{marker_list}{A list contains marker gene of cell type.}

\item{method}{Choosing the published method to assign cell type.}

\item{...}{other paramter used in method/classifer, the detial could see the corresponding method.}
}
\value{
The predict cell type.
}
\description{
Semi-supervised method to assign cell type
}
