% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{scclassifr}
\alias{scclassifr}
\title{scclassifr method}
\usage{
scclassifr(train, test, label_train, marker_list, time = F)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{marker_list}{cell type marker list contains marker genes.}

\item{time}{return computation time, default is FALSE.}
}
\value{
The predict cell type.
}
\description{
scclassifr method
}
