% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{scibet}
\alias{scibet}
\title{scibet method}
\usage{
scibet(train, test, label_train, norm = F)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{norm}{The data need normalize or not.}
}
\value{
The predict cell type.
}
\description{
scibet method
}
