% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{scid}
\alias{scid}
\title{scID method}
\usage{
scid(
  train,
  test,
  label_train,
  logFC = 0.6,
  only_pos = FALSE,
  estimate_weights_from_target = FALSE
)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{model}{Classification model supported via ‘caret’ package. A list of all models can be found here.}

\item{reclassify}{Cell types to reclassify using a different model}
}
\value{
The predict cell type.
}
\description{
scID method
}
