% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{scina}
\alias{scina}
\title{SCINA method}
\usage{
scina(
  test,
  marker_list,
  max_iter = 100,
  convergence_n = 10,
  convergence_rate = 0.999,
  sensitivity_cutoff = 0.9,
  rm_overlap = FALSE,
  allow_unknown = TRUE,
  log_file = "SCINA.log"
)
}
\arguments{
\item{test}{Same format like train.}

\item{marker_list}{A list contains marker gene of cell type.}

\item{max_iter}{An integer > 0. Default is 100. Max iterations allowed for the EM algorithm.}

\item{convergence_n}{An integer > 0. Default is 10. Stop the SCINA algorithm if during the last n rounds of iterations, cell type assignment keeps steady above the convergence_rate.}

\item{convergence_rate}{A float between 0 and 1. Default is 0.99. Percentage of cells for which the type assignment remains stable for the last n rounds.}

\item{sensitivity_cutoff}{A float between 0 and 1. Default is 1. The cutoff to remove signatures whose cells types are deemed as non-existent at all in the data by the SCINA algorithm.}

\item{rm_overlap}{A binary value, default 1 (TRUE), denotes that shared symbols between signature lists will be removed. If 0 (FALSE) then allows different cell types to share the same identifiers.}

\item{allow_unknown}{A binary value, default 1 (TRUE). If 0 (FALSE) then no cell will be assigned to the 'unknown' category.}

\item{log_file}{A string names the record of the running status of the SCINA algorithem, default 'SCINA.log'.}
}
\value{
The predict cell type.
}
\description{
SCINA method
}
