% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{sclearn}
\alias{sclearn}
\title{scLearn method}
\usage{
sclearn(
  train,
  test,
  label_train,
  species = "Mm",
  gene_low = 500,
  gene_high = 10000,
  mito_high = 0.1,
  min_cell_number = 10,
  umi_low = 1500,
  umi_high = Inf,
  logNormalize = TRUE,
  log_normalized = TRUE,
  threshold = 0.05,
  sample_information_timePoint = NULL,
  bootstrap_times = 10,
  cutoff = 0.01,
  dim_para = 0.999,
  vote_rate = 0.6,
  diff = 0.05,
  time = F
)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{species}{The species the cells belong to. Currently species "Hs" for homo sapiens or species "Mm" for mus musculus are available. It is used to detect mitochondrial genes, so species doesn't matter if data won't be considered the percentage of mitochondrial genes.}

\item{gene_low}{The minimum gene number for cells. Cells with genes below this threshold are filtered(default 500).}

\item{gene_high}{The maximum gene number for cells. cells with genes above this threshold are filtered(default 10000).}

\item{mito_high}{The maximum percentage of mitochondrial genes detected(default 0.1 among all the detected genes).}

\item{min_cell_number}{The minimal cell number to filter cell type. Default is 10.}

\item{umi_low}{The minimum number of unique molecular identifier induced (default 1500).}

\item{umi_high}{The maximum number of unique molecular identifier induced (default Inf).}

\item{logNormalize}{TRUE by default. If FALSE, the data will not be performed log normalized.}

\item{log_normalized}{Boolean, whether the train is log normalized (default : False).}

\item{threshold}{threshold to select feature, detail in M3Drop package.}

\item{sample_information_timePoint}{A character vector showing the time point of each sample. The column name of the vector is the sample name (default:NULL).}

\item{bootstrap_times}{The times for bootstrapping which should be at least larger than 1 (default:10).}

\item{cutoff}{The cutoff for selecting similarity threshold for each cell type (default:0.01).}

\item{dim_para}{The threshold to choose proper dimension for MDDM (default:0.999).}

\item{vote_rate}{Default is 0.6.}

\item{diff}{Default is 0.05.}

\item{time}{return computation time, default is FALSE.}
}
\value{
The predict cell type.
}
\description{
scLearn method
}
