% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{scmapcell}
\alias{scmapcell}
\title{------------------------
---Supervised methods---
------------------------
--------------------------
---UnSupervised methods---
--------------------------}
\usage{
scmapcell(train, test, label_train, w1 = 10, w2 = 3, threshold = 0.5)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{w1}{a positive integer specifying the number of nearest neighbours to find.}

\item{w2}{an integer specifying the number of nearest neighbours to find.}

\item{threshold}{the threshold which the maximum similarity between the query and a reference cell must exceed for the cell-type to be assigned.}
}
\value{
The predict cell type.
}
\description{
scmapcell method
}
