% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{scmapcluster}
\alias{scmapcluster}
\title{scmapcluster method}
\usage{
scmapcluster(train, test, label_train, threshold = 0.7, time = F)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{threshold}{threshold on similarity (or probability for SVM and RF)}

\item{time}{return computation time, default is FALSE.}
}
\value{
The predict cell type.
}
\description{
scmapcluster method
}
