% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{scsa}
\alias{scsa}
\title{SCSA method}
\usage{
scsa(test = NULL, marker_list = NULL, marker_file = NULL)
}
\arguments{
\item{test}{Same format like train.}

\item{marker_list}{A list contains marker gene of cell type.}

\item{marker_file}{The absolute path of marker file (now just support marker info from Seurat FindAllMarkers method).}
}
\value{
The predict cell type.
}
\description{
SCSA method
}
