% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{sctyper}
\alias{sctyper}
\title{sctyper method}
\usage{
sctyper(test, marker_list, cell.typing.method = "NTP", time = F)
}
\arguments{
\item{test}{Same format like train.}

\item{marker_list}{A list contains marker gene of cell type.}

\item{cell.typing.method}{cell typing method, c("NTP", "ES", "Average"), (default = "NTP").}

\item{time}{return computation time, default is FALSE.}
}
\value{
The predict cell type.
}
\description{
sctyper method
}
