% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{selfprojection}
\alias{selfprojection}
\title{Self projection of Supervised methods and Unsupervised methods}
\usage{
selfprojection(dataset, label_dataset, method, n = 5, ...)
}
\arguments{
\item{dataset}{A gene-cell matrix to do 5 fold cross validation.}

\item{method}{Choosing the Supervised/Unsupervised methods to assign cell type.}

\item{n}{Split vector to n-fold (deufault : 5).}

\item{...}{other paramter used in method/classifer, the detial could see the corresponding method.}

\item{label_train}{A vector of cell type label of train.}
}
\value{
The prediction of each methods.
}
\description{
Self projection of Supervised methods and Unsupervised methods
}
