% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{singlecellnet}
\alias{singlecellnet}
\title{singleCellNet method}
\usage{
singlecellnet(
  train,
  test,
  label_train,
  colName_samp = "row.names",
  nTopGenes = 10,
  nTopGenePairs = 25,
  nRand = 70,
  nTrees = 1000,
  nrand = 50
)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{colName_samp}{the name of the column that contains sample names (default : "row.names").}

\item{nTopGenes}{the number of classification genes per category (default : 10).}

\item{nTopGenePairs}{the number of top gene pairs per category (default : 25).}

\item{nRand}{number of random profiles generate for training (default : 70).}

\item{nTrees}{number of trees for random forest classifier (default : 1000).}

\item{nrand}{the number of random profiles generate for evaluation process (default : 50).}
}
\value{
The predict cell type.
}
\description{
singleCellNet method
}
