% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Supervised.R
\name{supervised}
\alias{supervised}
\title{Supervised method to assign cell type}
\usage{
supervised(train, test, label_train, method, ...)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{method}{Choosing the published method to assign cell type. Now support method contains: ACTINN, CaSTLe, CEHTAH, Garnett, SciBet, scID, scLearn, scmapcluster, scPred, scVI, Seurat, SingleCellNet, SingleR.}

\item{...}{other paramter used in method/classifer, the detial could see the corresponding method.}
}
\value{
The predict cell type.
}
\description{
Supervised method to assign cell type
}
