% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{actinn}
\alias{actinn}
\title{------------------------------
---Eager-supervised methods---
------------------------------}
\usage{
actinn(train, test, label_train, python_path, time = F)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{python_path}{Absolute python path that conda env is installed.}

\item{time}{return computation time, default is FALSE.}
}
\value{
The predict cell type.
}
\description{
ACTINN method
The tensorflow1 is used in ACTINN, please set the right python version in R through reticulate::use_python().
}
