% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{chetah}
\alias{chetah}
\title{CEHTAH method}
\usage{
chetah(train, test, label_train)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{threshold}{threshold for 'Unknown' cell type.}
}
\value{
The predict cell type.
}
\description{
CEHTAH method
}
