% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{digitalcellsorter}
\alias{digitalcellsorter}
\title{--------------------------
---UnSupervised methods---
--------------------------
----------------------------
---SemiSupervised methods---
----------------------------}
\usage{
digitalcellsorter(
  test,
  python_link = NULL,
  python_package_link = NULL,
  marker_list = NULL
)
}
\arguments{
\item{test}{Same format like train.}

\item{python_link}{The absolute path of python that DigitalCellSorter is installed, default is null.}

\item{python_package_link}{If DigitalCellSorter is installed and can not load, you can provide the absolute path of DigitalCellSorter package, like '/usr/local/lib/python3.8/site-packages/', default is null.}

\item{marker_list}{cell type marker list contains marker genes.}
}
\value{
The predict cell type.
}
\description{
DigitalCellSorter method
}
