% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{scid}
\alias{scid}
\title{scID method}
\usage{
scid(
  train,
  test,
  label_train,
  logFC = 0.6,
  only_pos = FALSE,
  estimate_weights_from_target = FALSE,
  time = F
)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{time}{return computation time, default is FALSE.}
}
\value{
The predict cell type.
}
\description{
scID method
}
