% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{scvi}
\alias{scvi}
\title{scVI method}
\usage{
scvi(train, test, label_train, n_epochs = 20, time = F)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{time}{return computation time, default is FALSE.}
}
\value{
The predict cell type.
}
\description{
scVI method
}
