% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{seurat}
\alias{seurat}
\title{Seurat method}
\usage{
seurat(train, test, label_train, k.filter = 200)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}
}
\value{
The predict cell type.
}
\description{
Seurat method
}
