% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteMethod.R
\name{singler}
\alias{singler}
\title{SingleR method}
\usage{
singler(
  train,
  test,
  label_train,
  method = "single",
  clusters = NULL,
  genes = "de",
  sd.thresh = 1,
  quantile = 0.8,
  fine.tune = TRUE,
  tune.thresh = 0.05,
  prune = TRUE,
  assay.type.test = "logcounts",
  assay.type.ref = "logcounts",
  check.missing = TRUE
)
}
\arguments{
\item{train}{A gene-cell matrix of single cell expression data.}

\item{test}{Same format like train.}

\item{label_train}{A vector of cell type label of train.}

\item{method}{Deprecated.}

\item{clusters}{A character vector or factor of cluster identities for each cell in \code{test}. If set, annotation is performed on the aggregated cluster profiles, otherwise it defaults to per-cell annotation.}

\item{genes, sd.thresh}{Arguments controlling the choice of marker genes used for annotation, see \code{\link{trainSingleR}}.}

\item{quantile, fine.tune, tune.thresh, prune}{Further arguments to pass to \code{\link{classifySingleR}}.}

\item{assay.type.test}{An integer scalar or string specifying the assay of \code{test} containing the relevant expression matrix, if \code{test} is a \linkS4class{SummarizedExperiment} object.}

\item{assay.type.ref}{An integer scalar or string specifying the assay of \code{ref} containing the relevant expression matrix, if \code{ref} is a \linkS4class{SummarizedExperiment} object (or is a list that contains one or more such objects).}

\item{check.missing}{Logical scalar indicating whether rows should be checked for missing values (and if found, removed).}
}
\value{
The predict cell type.
}
\description{
SingleR method
}
